+++ tests/vktestbinding.cpp
@@ -271,7 +271,7 @@ void Device::init(std::vector<const char *> &layers, std::vector<const char *> &
     const std::vector<VkQueueFamilyProperties> queue_props = phy_.queue_properties();
     std::vector<VkDeviceQueueCreateInfo> queue_info;
     queue_info.reserve(queue_props.size());
-    for (int i = 0; i < queue_props.size(); i++) {
+    for (uint32_t i = 0; i < (uint32_t)queue_props.size(); i++) {
         VkDeviceQueueCreateInfo qi = {};
         qi.sType = VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO;
         qi.pNext = NULL;
